+++ packages/thumbprint-react/components/Popover/test.tsx
@@ -4,14 +4,16 @@ import Popover from './index';
 
 // https://github.com/thumbtack/thumbprint/issues/72
 // https://github.com/FezVrasta/popper.js/issues/478#issuecomment-341494703
-jest.mock('popper.js', () => {
+jest.mock('popper.js', (): object => {
     const PopperJS = jest.requireActual('popper.js');
 
     class Popper {
-        constructor() {
+        public static placements: typeof PopperJS.placements;
+
+        public constructor() {
             return {
-                destroy: () => {},
-                scheduleUpdate: () => {},
+                destroy: (): void => {},
+                scheduleUpdate: (): void => {},
             };
         }
     }
@@ -21,10 +23,10 @@ jest.mock('popper.js', () => {
     return Popper;
 });
 
-test('renders a popover', () => {
+test('renders a popover', (): void => {
     const onCloseClick = jest.fn();
     const wrapper = mount(
-        <Popover isOpen onCloseClick={onCloseClick} launcher={() => null}>
+        <Popover isOpen onCloseClick={onCloseClick} launcher={(): null => null}>
             Goose
         </Popover>,
     );
@@ -32,12 +34,12 @@ test('renders a popover', () => {
     expect(wrapper).toMatchSnapshot();
 });
 
-test('initially traps focus to wrapper, not the focusable child', () => {
+test('initially traps focus to wrapper, not the focusable child', (): void => {
     jest.useFakeTimers();
 
     const onCloseClick = jest.fn();
     const wrapper = mount(
-        <Popover isOpen onCloseClick={onCloseClick} launcher={() => null}>
+        <Popover isOpen onCloseClick={onCloseClick} launcher={(): null => null}>
             <button type="button" id="hi">
                 Hi
             </button>